Add Virtualizer and WindowVirtualizer for Solid
