+++ src/solid/index.ts
@@ -3,3 +3,10 @@
  */
 export { VList } from "./VList";
 export type { VListProps, VListHandle } from "./VList";
+export { Virtualizer } from "./Virtualizer";
+export type { VirtualizerProps, VirtualizerHandle } from "./Virtualizer";
+export { WindowVirtualizer } from "./WindowVirtualizer";
+export type {
+  WindowVirtualizerProps,
+  // WindowVirtualizerHandle,
+} from "./WindowVirtualizer";
