+++ src/solid/utils.ts
@@ -1,0 +1,8 @@
+import { ItemsRange } from "../core/store";
+
+/**
+ * @internal
+ */
+export const isSameRange = (prev: ItemsRange, next: ItemsRange): boolean => {
+  return prev[0] === next[0] && prev[1] === next[1];
+};
